Avancement

+++ src/main/java/com/epf/API/dto/MapsDto.java
@@ -1,22 +1,22 @@
 package com.epf.API.dto;
 
 public class MapsDto {
-    private Long id_map;
+    private int id_map;
     private int ligne;
     private int colonne;
     private String chemin_image;
 
     public MapsDto() {}
 
-    public MapsDto(Long id_map, int ligne, int colonne, String chemin_image) {
+    public MapsDto(int id_map, int ligne, int colonne, String chemin_image) {
         this.id_map = id_map;
         this.ligne = ligne;
         this.colonne = colonne;
         this.chemin_image = chemin_image;
     }
 
-    public Long getId_map() { return id_map; }
-    public void setId_map(Long id_map) { this.id_map = id_map; }
+    public int getId_map() { return id_map; }
+    public void setId_map(int id_map) { this.id_map = id_map; }
 
     public int getLigne() { return ligne; }
     public void setLigne(int ligne) { this.ligne = ligne; }
