+++ src/main/java/com/epf/core/services/MapsServices.java
@@ -6,10 +6,10 @@ import java.util.Optional;
 
 public interface MapsServices {
     List<Maps> findAll();
-    Optional<Maps> findById(Long id);
+    Optional<Maps> findById(int id);
     Maps save(Maps map);
 
     void update(Maps plant);
 
-    void delete(Long id);
+    void delete(int id);
 }
