+++ src/main/java/com/epf/persistence/dao/PlantsDao.java
@@ -7,10 +7,10 @@ import java.util.Optional;
 
 public interface PlantsDao {
     List<Plants> findAll();
-    Optional<Plants> findById(Long id);
+    Optional<Plants> findById(int id);
     Plants save(Plants plant);
     void update(Plants plant);
-    void delete(Long id);
+    void delete(int id);
 }
